Abstract TryNode's decorator logic to a separate utility
